requirements.txt
# Core RAG
sentence-transformers==5.4.1
chromadb==1.5.8

# LLM
google-genai==1.73.1

# Data acquisition (only needed if you re-run scripts; safe to keep)
biopython==1.87

# UI
streamlit==1.57.0

# Utilities
python-dotenv==1.2.2
tqdm==4.67.3
pandas==3.0.2
numpy==2.4.4
PyYAML==6.0.3

# Note: torch is auto-installed as sentence-transformers dependency